Explicit memo support in raw image
Image is now returned and manipulated via an specific structure
to avoid the need to pass width or a separate pointer for memo
data.

KH-930 needs to be ported to the new image manipulation and
include explicit memo support. It's disabled for now.

Pattern decoding from memory also needs to gather explicit
memo information yet.

Enabled compilation debug for linux.